Various fixes for brushes
Remove global variables `_brushSize`, `_brushType`, `_brushColor`, and
`_brushTexture`, because global variables are evil.

The remove dependency of `Controls` on various brushes and their names.
This logic is moved to app.ts as it should be.

Now allow the user to adjust sizes of some more brushes.

Various other small fixes and tweaks.